Resolved problems marvin mentioned
